Added lock file and download file with callbacks
This update has a lock file to prevent users from uploading data at the
same time.  The lock file still needs to be integrated into the upload
function but creating the lock file and pushing and downloading it to
gdrive is functional.

It adds the users name, a timestamp (expires in 10 min), and a flag for
if the drive is actually locked or not.