Added bulk download and api calls
There should be enough api calls for everything assocaited with the
drive now.  We can upload/download all files including remotedb.json.

The lock system is now set up as well inside of the api calls.

TODO:  I did try to set up a system with the lock so it did not have to
re-download the lock file from the drive every time it needed to check
if the lock was active.  However, it does re-initiate the lock everytime
it runs the lock function. (ie. resets the timer).